refactor: transfer logic from hooks to components
