+++ src/features/forgotPassword/ForgotPasswordCard.tsx
@@ -1,34 +1,97 @@
 'use client'
+import { ChangeEvent, useRef, useState } from 'react'
+import ReCAPTCHA from 'react-google-recaptcha'
+import { SubmitHandler, useForm } from 'react-hook-form'
+
 import { Button } from '@/src/components/button/Button'
 import { Card } from '@/src/components/card/Card'
 import { Dialog } from '@/src/components/dialog/Dialog'
 import { Input } from '@/src/components/input/Input'
 import { Recaptcha } from '@/src/components/recaptcha/Recaptcha'
 import { Typography } from '@/src/components/typography/Typography'
-import { useForgotPasswordCard } from '@/src/features/forgotPassword/hooks/useForgotPasswordCard'
+import { FormType, schema } from '@/src/features/forgotPassword/validators'
+import { usePasswordRecoveryMutation } from '@/src/store/services/authApi'
+import { CustomerError } from '@/src/store/services/types'
+import { zodResolver } from '@hookform/resolvers/zod'
 import Link from 'next/link'
 
 import s from './forgotPassword.module.scss'
 
 export default function ForgotPasswordCard() {
+  const [isModalOpen, setIsModalOpen] = useState<boolean>(false)
+  const [formSubmit, setFormSubmit] = useState<boolean>(false)
+  const [passwordRecovery, { isLoading }] = usePasswordRecoveryMutation()
+  const [recaptchaError, setRecaptchaError] = useState<null | string>(null)
+  const [serverError, setServerError] = useState<null | string>(null)
+  const recaptchaRef = useRef<ReCAPTCHA | null>(null)
   const {
-    errors,
-    formSubmit,
+    clearErrors,
+    formState: { errors, isValid },
     getValues,
-    handleResendLink,
     handleSubmit,
-    isButtonDisabled,
-    isModalOpen,
-    onChangeEmail,
-    onChangeToken,
-    onSubmit,
-    recaptchaError,
-    recaptchaRef,
     register,
-    serverError,
-    setIsModalOpen,
+    reset,
+    setValue,
     trigger,
-  } = useForgotPasswordCard()
+  } = useForm<FormType>({
+    defaultValues: {
+      email: '',
+      recaptcha: '',
+    },
+    mode: 'onChange',
+    resolver: zodResolver(schema),
+  })
+
+  const onChangeToken = (value: null | string) => {
+    if (value === null) {
+      setRecaptchaError('Please verify that you are not a robot')
+      if (recaptchaRef.current) {
+        recaptchaRef.current.reset()
+      }
+    } else {
+      setValue('recaptcha', value)
+      setRecaptchaError(null)
+    }
+  }
+  const onChangeEmail = (event: ChangeEvent<HTMLInputElement>) => {
+    const value = event.target.value
+
+    setValue('email', value)
+    if (value) {
+      clearErrors('email')
+      setServerError(null)
+    }
+  }
+
+  const onSubmit: SubmitHandler<FormType> = async data => {
+    try {
+      await passwordRecovery({
+        baseUrl: process.env.NEXT_PUBLIC_BASE_URL as string,
+        email: data.email,
+        recaptcha: data.recaptcha,
+      }).unwrap()
+
+      setIsModalOpen(true)
+      setServerError(null)
+      clearErrors(['email', 'recaptcha'])
+      setFormSubmit(true)
+    } catch (error) {
+      const typedError = error as CustomerError
+
+      if (typedError?.data?.messages && typedError.data.messages[0]?.message) {
+        setServerError(typedError.data.messages[0].message)
+      } else {
+        setServerError("User with this email doesn't exist")
+      }
+    }
+  }
+
+  const handleResendLink = () => {
+    reset()
+    setFormSubmit(false)
+  }
+
+  const isButtonDisabled = !isValid || recaptchaError !== null || isLoading
 
   return (
     <div className={s.container}>
