+++ src/features/registration-confirmation-form/RegistrationConfirmationForm.tsx
@@ -1,15 +1,40 @@
 'use client'
+import { useEffect } from 'react'
+
 import { Button } from '@/src/components/button/Button'
 import { Typography } from '@/src/components/typography/Typography'
 import { AuthRoutes } from '@/src/constants/routing'
-import { useRegistrationConfirmationForm } from '@/src/features/registration-confirmation-form/hooks/useRegistrationConfirmationForm'
+import { useRegistrationConfirmationMutation } from '@/src/store/services/authApi'
+import { CustomerError } from '@/src/store/services/types'
 import Image from 'next/image'
 import Link from 'next/link'
+import { useRouter, useSearchParams } from 'next/navigation'
 
 import s from '@/src/features/registration-confirmation-form/email-confirmed.module.scss'
 
 export const RegistrationConfirmationForm = () => {
-  useRegistrationConfirmationForm()
+  const searchParams = useSearchParams()
+  const router = useRouter()
+  const [confirmRegistration] = useRegistrationConfirmationMutation()
+
+  useEffect(() => {
+    ;(async () => {
+      try {
+        // Получаем параметры через useSearchParams
+        const code = searchParams.get('code') as string
+
+        await confirmRegistration({ confirmationCode: code }).unwrap()
+      } catch (err) {
+        const error = err as CustomerError
+        const errorMessage = error.data?.messages[0]?.message
+
+        // Проверяем сообщение об ошибке и выполняем перенаправление при необходимости
+        if (errorMessage === 'Confirmation code is invalid') {
+          router.push('/auth/registration-email-resending')
+        }
+      }
+    })()
+  }, [searchParams, confirmRegistration, router])
 
   return (
     <div className={s.container}>
